requirements.txt
asgiref==3.8.1
behave==1.2.6
colorama==0.4.6
coverage==7.6.0
Django==5.0.7
django-js-asset==2.2.0
django-mptt==0.16.0
django-role-permissions==3.2.0
django_roles==0.0.4
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
parse==1.20.2
parse-type==0.6.2
pluggy==1.5.0
pytest==8.3.2
pytest-bdd==7.2.0
pytest-django==4.8.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
